feat(config) change Kong headers configuration
Instead of introducing a config option for each
and every header or set of headers, an array of
these values can be now specified using the `headers`
config option.

Only headers or tokens specified in the headers will
be set by Kong when applicable.

The goal here is to move towards a  simpler and
easier to understand configuration, similar to
1b9976f60 (#3147).